--- app/src/main/java/com/example/android/playlistmaker/data/NetworkClient.kt
@@ -1,7 +1,0 @@
-package com.example.android.playlistmaker.data
-
-import com.example.android.playlistmaker.data.dto.Response
-
-interface NetworkClient {
-    fun doRequest(dto: Any): Response
-}
