--- app/src/main/java/com/example/android/playlistmaker/data/dto/ITunesTrackRequest.kt
@@ -1,3 +1,0 @@
-package com.example.android.playlistmaker.data.dto
-
-data class ITunesTrackRequest(val term: String, val entity: String = "song")
