--- app/src/main/java/com/example/android/playlistmaker/data/dto/ITunesTrackResponse.kt
@@ -1,5 +1,0 @@
-package com.example.android.playlistmaker.data.dto
-
-import com.google.gson.annotations.SerializedName
-
-class ITunesTrackResponse(@SerializedName("results") val tracks: List<TrackDto>) : Response()
