--- app/src/main/java/com/example/android/playlistmaker/data/dto/StateDto.kt
@@ -1,5 +1,0 @@
-package com.example.android.playlistmaker.data.dto
-
-enum class StateDto {
-    STATE_DEFAULT, STATE_PLAYING, STATE_PAUSED, STATE_PREPARED
-}
