--- app/src/main/java/com/example/android/playlistmaker/domain/api/TracksRepository.kt
@@ -1,7 +1,0 @@
-package com.example.android.playlistmaker.domain.api
-
-import com.example.android.playlistmaker.domain.models.Track
-
-interface TracksRepository {
-    fun searchTracks(term: String): List<Track>
-}
