--- app/src/main/java/com/example/android/playlistmaker/domain/impl/AudioPlayerInteractorImpl.kt
@@ -1,36 +1,0 @@
-package com.example.android.playlistmaker.domain.impl
-
-import com.example.android.playlistmaker.domain.api.AudioPlayerInteractor
-import com.example.android.playlistmaker.domain.api.AudioPlayerRepository
-import com.example.android.playlistmaker.domain.models.State
-
-class AudioPlayerInteractorImpl(private val player: AudioPlayerRepository) : AudioPlayerInteractor {
-    override fun startPlayer() {
-        player.startPlayer()
-    }
-
-    override fun pausePlayer() {
-        player.pausePlayer()
-    }
-
-    override fun playbackControl(): State {
-        return player.playbackControl()
-    }
-
-    override fun preparePlayer(
-        trackUrl: String,
-        onPrepared: () -> Unit,
-        onCompletePlay: () -> Unit,
-    ) {
-        player.preparePlayer(trackUrl, onPrepared, onCompletePlay)
-    }
-
-    override fun release() {
-        player.release()
-    }
-
-    override fun getTrackTime(): Int {
-        return player.currentPosition
-    }
-
-}
