--- app/src/main/java/com/example/android/playlistmaker/domain/models/State.kt
@@ -1,5 +1,0 @@
-package com.example.android.playlistmaker.domain.models
-
-enum class State {
-    STATE_DEFAULT, STATE_PREPARED, STATE_PLAYING, STATE_PAUSED
-}
