--- app/src/main/java/com/example/android/playlistmaker/domain/models/Track.kt
@@ -1,19 +1,0 @@
-package com.example.android.playlistmaker.domain.models
-
-import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Track(
-    val trackId: Long,
-    val trackName: String,
-    val artistName: String,
-    @SerializedName("trackTimeMillis") val trackTime: Long,
-    val artworkUrl100: String,
-    val collectionName: String,
-    val primaryGenreName: String,
-    val releaseDate: String,
-    val country: String,
-    val previewUrl: String?,
-) : Parcelable
