+++ app/src/main/java/com/example/android/playlistmaker/medialibrary/ui/MediaLibraryActivity.kt
@@ -1,0 +1,12 @@
+package com.example.android.playlistmaker.medialibrary.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.android.playlistmaker.R
+
+class MediaLibraryActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_media_library)
+    }
+}
