+++ app/src/main/java/com/example/android/playlistmaker/player/data/dto/StateDto.kt
@@ -1,0 +1,5 @@
+package com.example.android.playlistmaker.player.data.dto
+
+enum class StateDto {
+    STATE_DEFAULT, STATE_PLAYING, STATE_PAUSED, STATE_PREPARED
+}
