+++ app/src/main/java/com/example/android/playlistmaker/player/domain/api/AudioPlayerInteractor.kt
@@ -1,0 +1,17 @@
+package com.example.android.playlistmaker.player.domain.api
+
+import com.example.android.playlistmaker.player.domain.models.State
+
+interface AudioPlayerInteractor {
+    fun startPlayer()
+
+    fun pausePlayer()
+
+    fun playbackControl(): State
+
+    fun preparePlayer(trackUrl: String, onPrepared: () -> Unit, onCompletePlay: () -> Unit)
+
+    fun release()
+
+    fun getTrackTime(): Int
+}
