+++ app/src/main/java/com/example/android/playlistmaker/player/domain/models/State.kt
@@ -1,0 +1,5 @@
+package com.example.android.playlistmaker.player.domain.models
+
+enum class State {
+    STATE_DEFAULT, STATE_PREPARED, STATE_PLAYING, STATE_PAUSED
+}
