--- app/src/main/java/com/example/android/playlistmaker/presentation/ui/medialibrary/MediaLibraryActivity.kt
@@ -1,12 +1,0 @@
-package com.example.android.playlistmaker.presentation.ui.medialibrary
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.android.playlistmaker.R
-
-class MediaLibraryActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_media_library)
-    }
-}
