--- app/src/main/java/com/example/android/playlistmaker/presentation/ui/search/TrackViewHolder.kt
@@ -1,44 +1,0 @@
-package com.example.android.playlistmaker.presentation.ui.search
-
-import android.util.TypedValue
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.example.android.playlistmaker.R
-import com.example.android.playlistmaker.domain.models.Track
-import java.text.SimpleDateFormat
-import java.util.Locale
-
-class TrackViewHolder(parent: ViewGroup) : RecyclerView.ViewHolder(
-    LayoutInflater.from(parent.context).inflate(R.layout.track_element, parent, false)
-) {
-
-    private val trackName: TextView = itemView.findViewById(R.id.tvTrackName)
-    private val artistName: TextView = itemView.findViewById(R.id.tvArtistName)
-    private val trackTime: TextView = itemView.findViewById(R.id.tvTrackDuration)
-
-    fun bind(model: Track) {
-        trackName.text = model.trackName
-        artistName.text = model.artistName
-        trackTime.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(model.trackTime)
-        val urlCover = itemView.findViewById<ImageView>(R.id.ivTrack)
-        Glide.with(itemView.context).load(model.artworkUrl100).placeholder(R.drawable.placeholder)
-            .centerInside().transform(
-                RoundedCorners(
-                    TypedValue.applyDimension(
-                        TypedValue.COMPLEX_UNIT_DIP,
-                        ROUND_CORNERS_SIZE_PX,
-                        itemView.context.resources.displayMetrics
-                    ).toInt()
-                )
-            ).into(urlCover)
-    }
-
-    companion object {
-        private const val ROUND_CORNERS_SIZE_PX = 2f
-    }
-}
