+++ app/src/main/java/com/example/android/playlistmaker/search/data/dto/ITunesTrackRequest.kt
@@ -1,0 +1,3 @@
+package com.example.android.playlistmaker.search.data.dto
+
+data class ITunesTrackRequest(val term: String, val entity: String = "song")
