+++ app/src/main/java/com/example/android/playlistmaker/search/data/dto/ITunesTrackResponse.kt
@@ -1,0 +1,5 @@
+package com.example.android.playlistmaker.search.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+class ITunesTrackResponse(@SerializedName("results") val tracks: List<TrackDto>) : Response()
