+++ app/src/main/java/com/example/android/playlistmaker/search/domain/api/TracksInteractor.kt
@@ -1,0 +1,12 @@
+package com.example.android.playlistmaker.search.domain.api
+
+import com.example.android.playlistmaker.search.domain.models.Track
+
+interface TracksInteractor {
+
+    fun searchTracks(term: String, consumer: TracksConsumer)
+
+    interface TracksConsumer {
+        fun consume(foundTracks: List<Track>)
+    }
+}
