+++ app/src/main/java/com/example/android/playlistmaker/search/domain/api/TracksRepository.kt
@@ -1,0 +1,7 @@
+package com.example.android.playlistmaker.search.domain.api
+
+import com.example.android.playlistmaker.search.domain.models.Track
+
+interface TracksRepository {
+    fun searchTracks(term: String): List<Track>
+}
