+++ app/src/main/java/com/example/android/playlistmaker/search/domain/impl/TracksInteractorImpl.kt
@@ -1,0 +1,16 @@
+package com.example.android.playlistmaker.search.domain.impl
+
+import com.example.android.playlistmaker.search.domain.api.TracksInteractor
+import com.example.android.playlistmaker.search.domain.api.TracksRepository
+import java.util.concurrent.Executors
+
+class TracksInteractorImpl(private val respository: TracksRepository) : TracksInteractor {
+
+    private val executor = Executors.newCachedThreadPool()
+
+    override fun searchTracks(term: String, consumer: TracksInteractor.TracksConsumer) {
+        executor.execute {
+            consumer.consume(respository.searchTracks(term))
+        }
+    }
+}
