+++ app/src/main/res/layout/activity_audio_player.xml
@@ -9,7 +9,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".presentation.ui.audioplayer.AudioPlayerActivity">
+        tools:context=".player.ui.AudioPlayerActivity">
 
         <FrameLayout
             android:id="@+id/flBackToMain"
